+++ bin/ast.dart
@@ -4,68 +4,109 @@ import 'package:analyzer/analyzer.dart';
 import 'package:analyzer/src/generated/ast.dart';
 import 'package:analyzer/src/generated/scanner.dart';
 
-bool isSubType(dynamic obj, ClassMirror sec) =>
-    reflect(obj).type.isSubtypeOf(sec);
+bool isSubType(dynamic obj, List<ClassMirror> sec) {
+  for (ClassMirror type in sec) {
+    if (noSubTypes) {
+      if (reflect(obj).type.reflectedType.toString() ==
+          type.reflectedType.toString()) {
+        return true;
+      }
+    } else {
+      if (reflect(obj).type.isSubtypeOf(type)) {
+        return true;
+      }
+    }
+  }
+  return false;
+}
 
 Type typeOf(dynamic obj) => reflect(obj).type.reflectedType;
+
+bool noSubTypes = false;
+List<String> failures = [],
+    stuff = [];
 
 main(List<String> args) {
   if (args.length == 1) {
     CompilationUnit unit = parseDartFile(args[0]);
-    ClassMirror mirror = reflectClass(Object);
-    print(expandedAst(unit, mirror));
-  } else if (args.length == 2) {
+    List<ClassMirror> mirrors = [reflectClass(Object)];
+    print(expandedAst(unit, mirrors));
+  } else if (args.length >= 2) {
+    List<String> typeArgs = args.getRange(1, args.length).toList();
+    if (typeArgs.last == '-n') {
+      typeArgs.removeLast();
+      noSubTypes = true;
+    }
     CompilationUnit unit = parseDartFile(args[0]);
-    List mirrors = createMirrors(args[1]);
-    if (mirrors[0] != null) {
-      print(expandedAst(unit, mirrors[0]));
-    } else if (mirrors[1] != null) {
-      print(expandedAst(unit, mirrors[1]));
+    List<ClassMirror> mirrors = createMirrors(typeArgs);
+    if (mirrors.isNotEmpty) {
+      print(expandedAst(unit, mirrors));
     } else {
-      print('Invalid AST node Type: ' + args[1]);
+      print('Invalid AST node Types: ' + typeArgs.join(' '));
       exit(2);
     }
   } else {
     print('''
-        Usage: ast <path> <Type(optional)> ||| prints the whole expanded AST
+        Usage: ast <path> <Type(optional)> <-n(optional)>
 
         <path> must be absolute
         <Type> will print all nodes that are subtypes of Type
+        Can enter multiple <Type> arguments.
+        To print Types without their subtypes append -n.
         ''');
     exit(2);
   }
 }
 
-List createMirrors(String arg) {
-  MirrorSystem mirrors = currentMirrorSystem();
-  LibraryMirror engineAst = mirrors.libraries.values.firstWhere(
-      (LibraryMirror engineAst) =>
-          engineAst.qualifiedName == new Symbol('engine.ast'));
-  LibraryMirror engineScanner = mirrors.libraries.values.firstWhere(
-      (LibraryMirror engineScanner) =>
-          engineScanner.qualifiedName == new Symbol('engine.scanner'));
-  ClassMirror cmAst;
-  ClassMirror cmScanner;
-  try {
-    cmAst = engineAst.declarations[new Symbol(arg)];
-    cmScanner = engineScanner.declarations[new Symbol(arg)];
-  } catch (exception, stackTrace) {
-    //
+List<ClassMirror> createMirrors(List<String> types) {
+  List<Symbol> symbols = [];
+  for (var type in types) {
+    symbols.add(new Symbol(type));
   }
-  return [cmAst, cmScanner];
+  MirrorSystem mirrors = currentMirrorSystem();
+  Map<Symbol, DeclarationMirror> combinedEngine = {};
+  Map<Symbol, DeclarationMirror> engineAst = mirrors.libraries.values
+      .firstWhere((LibraryMirror engine) =>
+          engine.qualifiedName == new Symbol('engine.ast')).declarations;
+  Map<Symbol, DeclarationMirror> engineScanner = mirrors.libraries.values
+      .firstWhere((LibraryMirror engine) =>
+          engine.qualifiedName == new Symbol('engine.scanner')).declarations;
+  List<ClassMirror> cmAst = [];
+  combinedEngine
+    ..addAll(engineAst)
+    ..addAll(engineScanner);
+  for (var symbol in symbols) {
+    if (combinedEngine.containsKey(symbol)) {
+      cmAst.add(combinedEngine[symbol]);
+    } else {
+      //TODO: Replace with a proper regex
+      failures.add(symbol
+          .toString()
+          .replaceAll('Symbol(', '')
+          .replaceAll('(', '')
+          .replaceAll(')', ''));
+    }
+  }
+  return cmAst;
 }
 
-String expandedAst(CompilationUnit u, ClassMirror c) {
+String expandedAst(CompilationUnit u, List<ClassMirror> c) {
   Iterable children = u.childEntities;
   getRecursiveChildren(children, c);
   String temp = stuff.join('\n');
+  if (failures.isNotEmpty) {
+    temp += '\nInvalid AST node Types (not printed): ';
+    for (var item in failures) {
+      temp += '$item ';
+    }
+  }
   stuff.clear();
+  failures.clear();
+  noSubTypes = false;
   return temp;
 }
 
-List<String> stuff = [];
-
-void getRecursiveChildren(dynamic cu, ClassMirror nodeType,
+void getRecursiveChildren(dynamic cu, List<ClassMirror> nodeTypes,
     [List<String> builder, int count]) {
   if (builder == null) {
     builder = [];
@@ -73,7 +114,7 @@ void getRecursiveChildren(dynamic cu, ClassMirror nodeType,
   }
   for (var ch in cu) {
     var temp = null;
-    if (isSubType(ch, nodeType)) {
+    if (isSubType(ch, nodeTypes)) {
       count++;
       stuff.add(spaces(count) + typeOf(ch).toString() + '::: ' + ch.toString());
     }
@@ -83,25 +124,25 @@ void getRecursiveChildren(dynamic cu, ClassMirror nodeType,
       //
     }
     if (temp != null) {
-      if (isSubType(ch, nodeType)) {
+      if (isSubType(ch, nodeTypes)) {
         builder.add(stuff.toString());
-        getRecursiveChildren(temp, nodeType, builder, count);
+        getRecursiveChildren(temp, nodeTypes, builder, count);
         count--;
       } else {
         builder.add(stuff.toString());
-        getRecursiveChildren(temp, nodeType, builder, count);
+        getRecursiveChildren(temp, nodeTypes, builder, count);
       }
     }
   }
 }
 
 String spaces(int c) {
-  List<String> spac = [];
+  List<String> space = [];
   if (c > 1) {
     for (int i = 1; i < c; i++) {
-      spac.add('  ');
+      space.add('  ');
     }
-    return spac.join();
+    return space.join();
   } else {
     return '';
   }