+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
+## 0.0.4
+
+- Added optional argument ```-n```, "no subtypes".
+- Added support for multiple ```Type``` arguments.
+- More useful error messages.
+
 ## 0.0.3
 
 - Removed optional arguments ```-dec``` and ```-dir```.
 - Added optional argument ```Type```.
-- Better type printing.
+- Better type inference.
 
 ## 0.0.2